+++ Samples~/Scripts/CreateIns.cs
@@ -1,0 +1,12 @@
+namespace SaintsField.Samples.Scripts
+{
+    public class CreateIns : SaintsMonoBehaviour
+    {
+        [GetPrefabWithComponent, Expandable] public SaintsArray2DRBuildTest prefab;
+
+        private void Awake()
+        {
+            Instantiate(prefab, transform);
+        }
+    }
+}